requirements.txt
requests==2.25.1
python-telegram-bot==13.12